refactor: remove key `isServiceFactory`
1. Remove key `isServiceFactory`

This key does not seem required anymore. Without it the tests still pass
without any warning and a consuming application also does not get any
warning when using this add-on.

This key seems to be a relic from Ember.js <= v2, where it was required
when a Servive was extending `Ember.ObjectProxy` instead of `Service`.

See:
- https://github.com/DockYard/ember-cart/pull/21
- https://github.com/emberjs/ember.js/search?q=isServiceFactory&type=issues

2. Condense export of the Service

Using a `const` was already not necessary before, we could have chained
`reopenClass()` right after the `ObjectProxy.extend()`.