ci: run workflow only after lint & tests passed
Currently the workflows `CI` and `tag-release-publish` are executed in
parallel:

- Workflow `CI`
  https://github.com/peopledoc/ember-cli-embedded/runs/4020380098?check_suite_focus=true#step:1:1

- Workflow `tag-release-publish`
  https://github.com/peopledoc/ember-cli-embedded/runs/4020380101?check_suite_focus=true#step:1:1

By enabling the option `Show timestamps`, using the cog on the right, we
can see both started at the same time (+/- a few seconds of difference).

Which is wrong as we could potentially tag/release/publish while the
lint/tests failed. (this is unlikely to happen since the workflow `CI`
already runs for any pull request, but better safe than sorry right?)

---

This `if` condition is needed because:

https://docs.github.com/en/actions/learn-github-actions/events-that-trigger-workflows#workflow_run

> `workflow_run`
> (...)
> A workflow run is triggered regardless of the result of the previous workflow.