ci: fix concurrency group's name
The context `${{ github.head_ref }}` is only available in PR-related
events, see:

https://docs.github.com/en/actions/learn-github-actions/contexts#github-context

> The `head_ref` or source branch of the pull request in a workflow run.
> This property is only available when the event that triggers a
> workflow run is either `pull_request` or `pull_request_target`.

So when this worklow was running for the `master` branch, the name was
empty, hence the error spoted in this run:

https://github.com/peopledoc/ember-cli-embedded/actions/runs/1398906462

> The workflow is not valid. .github/workflows/ci.yml (Line: 13, Col: 10): Unexpected value ''

Now, the group's name will never be empty, with at least a prefix `ci-`.

Using the context `github.ref` might not even be necessary because:
- in PR events, `github.head_ref` will always return a value
- when running on `master` we would still have the prefix `ci-`

But this way we will have a nice name displayed such as `ci-master` etc.